Assume neutron port exists in VIF attach
When a VIF is attached to a portgroup with a MAC address, VIFPortIDMixin
updates the MAC address of the neutron port with the MAC address of the
portgroup.  The common mixin class VIFPortIDMixin is currently designed
to support non-neutron network interfaces. In order to do this, the
neutron port is first queried, and this query is allowed to fail. This
behaviour is both inefficient, requiring an unnecessary neutron API
call; and error prone, as failures are ignored when accessing the
neutron API.

This change removes the neutron port query during VIF attachment, and
assumes that the VIF being attached does in fact reference a valid
neutron port.

Change-Id: Iab1a3a1ee1f617f09b54183f6ba786cf2ee1f62d