Remove duplicate server-ip-address DHCP option
This option was added by name in addition to the numeric form. When
reverting to the named options we ended up with the named option twice,
which broke the API request validation in neutron.

To compound the issue, the API request validation code in neutron does
not handle non-hashable types, including the dicts used to define the
DHCP options. This led to a cryptic error such as:

PUT failed.: TypeError: unhashable type: 'dict'

This change removes the duplicate option.

Change-Id: I9b5c646dcbd4a76ff62fc898dc225100abaabcfa