WIP: Decompose core deploy step
Currently all almost all of the deployment process is contained in a
single deploy step called deploy, on the deploy interface. This
restricts the customisation that can be applied via deploy steps, since
steps may only be added before or after this step.

This change decomposes the current deploy step into multiple steps on
the deploy interface:

* deploy (100)
* continue_deploy (90)
* deploy_is_done ('direct' interface only) (85)
* reboot_to_instance (80)

Backwards compatibility with drivers that do not support decomposed
steps is maintained via a 'has_decomposed_deploy_steps' method on the
agent heartbeat mixin.

Change-Id: Ief586473aca0e22b74efe83ef70c354fd5df17bf
Story: 2006963
Task: 37778