Make port binding failure (configurably) fatal
When using drivers like `networking-generic-switch`, there can be issues
with changing the VLAN the server is attached to.

Currently, the instance will still continue to attempt to boot, in some
cases the instance can go into ACTIVE (if the server happens to be on
the provisioning network already), but the server is not attached to the
correct VLAN, generally causing the network to fail to come up.

If we look at the ports, they are marked as binding failed. This is
something that Nova checks for after doing the port update to attach the
binding information. Lets add that check into Ironic, such that the
above build will fail with a good hint at where to look for problems
(namely Neutron).

Change-Id: I3811941a3dff0a9f968258d05cc020e1f52e3e40